requirements.txt
fastapi
uvicorn
jinja2
python-multipart
supabase
python-dotenv
httpx